src/requirements.txt
Flask==2.2.3
flask_socketio==5.3.4
numpy==1.19.1
paho_mqtt==1.6.1
torch==2.0.1
ultralytics==8.0.120
